add object level API instead of only DFS API to allow the split to be done  for regular KV object.
